make lookup table threshold a floating point value
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@32874 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/algo/blast/api/blast_options.hpp
@@ -123,10 +123,10 @@ public:
 
     /******************* Lookup table options ***********************/
     /// Returns WordThreshold
-    int GetWordThreshold() const;
+    double GetWordThreshold() const;
     /// Sets WordThreshold
     /// @param w WordThreshold [in]
-    void SetWordThreshold(int w);
+    void SetWordThreshold(double w);
 
 #ifndef SKIP_DOXYGEN_PROCESSING
